ui-charts: create basicOnPan in useSyncManchette
